Add files via upload
Brain has billions of neurons that each connect to a lot of others. They communicate with each
other via electric currents, i.e., brain waves. They are oscillating electrical charges measuring
just a few millionths of volts and are propagated along the perineural system and blend with
this continuous current. This activity is actually really strong. Enough to be detected on the
outside by placing electrodes on the scalp. This process can be called EEG.
EEG simply means electric brain graph and can be used to evaluate the electrical activities
in the brain. One way it works is by conveying information in their rate of repetition. Some
oscillations, measured on the scalp, occur at more than 30 cycles/frequencies per second. This
can be used to categorize them in terms of their level of activity and frequencies each of which
corresponds to different states of thoughts or experience. Namely, Delta(0.5 - 4 Hz), Theta(4 -
7 Hz), Alpha(8 - 13 z), Beta(13 - 30 z), Gamma(>= 35 Hz).